Pull out canister method building
