+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_ast/mod.rs
@@ -5,6 +5,7 @@ use swc_ecma_parser::{lexer::Lexer, Parser, StringInput, Syntax, TsConfig};
 
 use crate::ts_ast::{source_map::SourceMapped, AzleProgram};
 
+mod build_canister_methods;
 mod build_datatype_nodes;
 mod build_query_methods;
 mod build_update_methods;
